+++ packages/runtime/src/widgets/Form/Dropdown.tsx
@@ -27,6 +27,8 @@ import type { HasBorder } from "../../shared/hasSchema";
 import type { FormInputProps } from "./types";
 import { EnsembleFormItem } from "./FormItem";
 
+const widgetName = "Dropdown";
+
 export type DropdownStyles = {
   dropdownBackgroundColor?: string;
   dropdownBorderRadius?: number;
@@ -61,7 +63,7 @@ const Dropdown: React.FC<DropdownProps> = (props) => {
   >();
   const { "item-template": itemTemplate, ...rest } = props;
   const { id, rootRef, values } = useRegisterBindings(
-    { ...rest, initialValue: props.value, selectedValue },
+    { ...rest, initialValue: props.value, selectedValue, widgetName },
     props.id,
     {
       setSelectedValue,
@@ -264,4 +266,4 @@ const Dropdown: React.FC<DropdownProps> = (props) => {
   );
 };
 
-WidgetRegistry.register("Dropdown", Dropdown);
+WidgetRegistry.register(widgetName, Dropdown);