+++ packages/runtime/src/widgets/Form/Form.tsx
@@ -14,6 +14,8 @@ import type {
   EnsembleWidgetStyles,
 } from "../../shared/types";
 
+const widgetName = "Form";
+
 export type FormProps = {
   children: EnsembleWidget[];
   enabled: boolean;
@@ -25,7 +27,9 @@ export type FormProps = {
 } & EnsembleWidgetProps;
 export const Form: React.FC<FormProps> = (props) => {
   const { children, ...rest } = props;
-  const { values } = useRegisterBindings({ ...rest });
+
+  const [form] = AntForm.useForm();
+  const getValues = form.getFieldsValue;
 
   const action = useEnsembleAction(props.onSubmit);
   const onFinishCallback = useCallback(
@@ -39,9 +43,12 @@ export const Form: React.FC<FormProps> = (props) => {
     [action],
   );
 
+  const { values } = useRegisterBindings({ ...rest }, rest.id, { getValues });
+
   return (
     <AntForm
       colon={false}
+      form={form}
       layout={getLayout(values?.styles?.labelPosition)}
       onFinish={onFinishCallback}
       style={{
@@ -68,4 +75,4 @@ const getLayout = (labelPosition?: string): FormLayout => {
   }
 };
 
-WidgetRegistry.register("Form", Form);
+WidgetRegistry.register(widgetName, Form);